requirements.txt
echoÿ"Flask==2.3.3
openai==1.3.3
python-dotenv==1.0.0" > requirements.txt
